requirements.txt
opencv-python
cryptography
patool